chore(week-4): formatted contracts

+++ src/week_4/contracts/ArbitraryAlert.sol
@@ -13,13 +13,13 @@ contract Hero {
 
     struct Ambush {
         bool alerted;
-        uint enemies;
+        uint256 enemies;
         bool armed;
     }
 
-    uint lastContact;
+    uint256 lastContact;
 
-    function alert(uint enemies, bool armed) external {
+    function alert(uint256 enemies, bool armed) external {
         ambush = Ambush(true, enemies, armed);
     }
 }
