+++ src/week_4/contracts/Getter.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.4;
 
 contract Getter {
-	uint public value;
+    uint256 public value;
 
-	constructor(uint _value) {
-		value = _value;
-	}
+    constructor(uint256 _value) {
+        value = _value;
+    }
 }
