+++ src/week_4/contracts/PureDouble.sol
@@ -2,11 +2,11 @@
 pragma solidity ^0.8.4;
 
 contract PureDouble {
-    function double(uint num) external pure returns(uint){
+    function double(uint256 num) external pure returns (uint256) {
         return num * 2;
     }
 
-    function double(uint num1, uint num2) external pure returns(uint,uint){
-        return (num1*2,num2*2);
+    function double(uint256 num1, uint256 num2) external pure returns (uint256, uint256) {
+        return (num1 * 2, num2 * 2);
     }
 }
