+++ src/week_4/contracts/Setter.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.4;
 
 contract Setter {
-	uint public value;
+    uint256 public value;
 
-	function modify(uint _value) external {
-		value = _value;
-	}
+    function modify(uint256 _value) external {
+        value = _value;
+    }
 }
