+++ src/week_4/contracts/SumAndAverage.sol
@@ -2,9 +2,8 @@
 pragma solidity ^0.8.4;
 
 contract SumAndAverage {
-    
-    function sumAndAverage(uint a,uint b,uint c, uint d) external pure returns(uint,uint){
-        uint sum = a + b + c +d;
+    function sumAndAverage(uint256 a, uint256 b, uint256 c, uint256 d) external pure returns (uint256, uint256) {
+        uint256 sum = a + b + c + d;
         return (sum, sum / 4);
     }
 }
