+++ src/week_4/contracts/Transfer.sol
@@ -2,13 +2,13 @@
 pragma solidity ^0.8.4;
 
 contract Transfer {
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
     constructor() {
         balances[msg.sender] = 1000;
     }
 
-    function transfer(address beneficiary, uint amount) external {
+    function transfer(address beneficiary, uint256 amount) external {
         require(balances[msg.sender] >= amount, "Balance too low!");
         balances[beneficiary] += amount;
         balances[msg.sender] -= amount;
