+++ src/week_4/contracts/WithSignature.sol
@@ -2,12 +2,10 @@
 pragma solidity ^0.8.4;
 
 contract Sidekick {
-    function sendAlert(address hero, uint enemies, bool armed) external {
-        bytes memory payload = abi.encodeWithSignature("alert(uint256,bool)",enemies,armed);
+    function sendAlert(address hero, uint256 enemies, bool armed) external {
+        bytes memory payload = abi.encodeWithSignature("alert(uint256,bool)", enemies, armed);
 
-        (bool success, ) = hero.call(
-            payload
-        );
+        (bool success,) = hero.call(payload);
 
         require(success);
     }
@@ -18,11 +16,11 @@ contract Hero {
 
     struct Ambush {
         bool alerted;
-        uint enemies;
+        uint256 enemies;
         bool armed;
     }
 
-    function alert(uint enemies, bool armed) external {
+    function alert(uint256 enemies, bool armed) external {
         ambush = Ambush(true, enemies, armed);
     }
 }
